refactor: renombrar entidad New a SavedNew y servicios asociados
