+++ aspnet-core/src/Newslify.Application/NewslifyApplicationAutoMapperProfile.cs
@@ -2,7 +2,7 @@
 using Newslify.Keywords;
 using Newslify.Languages;
 using Newslify.LogReadNews;
-using Newslify.News;
+using Newslify.SavedNews;
 using Newslify.ReadingLists;
 
 namespace Newslify;
@@ -15,7 +15,7 @@ public class NewslifyApplicationAutoMapperProfile : Profile
          * Alternatively, you can split your mapping configurations
          * into multiple profile classes for a better organization. */
         CreateMap<Language, LanguageDto>();
-        CreateMap<New, NewDto>();
+        CreateMap<SavedNew, SavedNewDto>();
         CreateMap<Keyword, KeywordDto>();
         CreateMap<LogReadNew, LogReadNewDto>();
         CreateMap<ReadingList, ReadingListDto>();
